refactor: move 'addressSummary' to type WalletInfo
